requirements.txt
torch
torchvision
torchaudio
opencv-python
pillow
matplotlib
